// src/app/models/intimation.ts
export class Intimation 	{
    IntimationID: number;
    IntimationStatus: number;
    AgentMobileNo?: string;
    AgentCode?: string;
    AgentName?: string;
    AgentEmail?: string;
    ROOffice?: number;
    InsurerName?: number;
    DivisionName?: number;
    BranchName?: number;
    InsurerRefNo?: string;
    InsuredAddress?: string;
    ContactPersonMobileNo?: string;
    ManufactureYear?: number;
    VehicleType?: number;
    Model?: number;
    StartingLocationFrom?: string;
    InspectionLocationTo?: string;
    Kgm?: string;
    Remarks?: string;
    Pincode?: string;
    InsuredName?: string;
    InsuredMobileNo?: string;
    VehicleRegistrationNo?: string;
    InspectionType?: number;
    Manufacturer?: number;
    Variant?: number;
    PaymentMode?: number;
    IntimationDate?: string;
    CCE?: string;
    Surveyor?: number;
    QCHoldReason?: number;
    StepStatus?: number;
    IntimationRemark?: string;
    ExtraKm?: string;
    AppointmentOn?: any;
    FourWheelerType?: number;
    EngineNo?: string;
    ChassisNo?: string;
    UpdateStatus?: number;
    ExtraKillometer?: string;
    ManualReportNo?: string;
    UnderTakingSlipNo?: string;
    ReceiptNo?: string;
    VehicleLocation?: any;
    CngLpgFitted?: any;
    CngLpgEndorsdinRC?: any;
    FrontNoPlate?: any;
    RearNoPlate?: any;
    FrontBumper?: any;
    Grill?: any;
    HeadLampLT?: any;
    HeadLampRT?: any;
    IndicatorLightLT?: any;
    IndicatorLightRT?: any;
    FogLampLT?: any;
    FogLampRT?: any;
    FrontPanel?: any;
    Bonnet?: any;
    LeftApron?: any;
    RightApron?: any;
    LTFrontFender?: any;
    LTFrontDoor?: any;
    LTRearDoor?: any;
    LTPillarDoorA?: any;
    LTPillarCenterB?: any;
    LTPillarRearC?: any;
    LTRunningBoard?: any;
    LTQtrPanel?: any;
    Dicky?: any;
    RearBumper?: any;
    TailLampLT?: any;
    TailLampRT?: any;
    RTQtrPanel?: any;
    RTRearDoor?: any;
    RTFrontDoor?: any;
    RTPillarFrontA?: any;
    RTCenterPillarB?: any;
    RTPillarRearC?: any;
    RTRunningBoard?: any;
    RTFrontFender?: any;
    FloorSilencer?: any;
    UnderCarriage?: any;
    RearViewMirrorLT?: any;
    RearViewMirrorRT?: any;
    Rim?: any;
    Colour?: any;
    RCVerified?: any;
    FuelUsed?: any;
    StereoMake?: any;
    CDChangerMake?: any;
    AnyOther?: any;
    SeatCovers?: any;
    CentreLock?: any;
    GearLocking?: any;
    FrontwsGlassLaminated?: any;
    LeftFrontDoorGlass?: any;
    LeftRearDoorGlass?: any;
    RightFrontDoorGlass?: any;
    RightRearDoorGlass?: any;
    BackGlass?: any;
    Tyres?: any;
    RHFrontTyreMake?: any;
    RHFrontTyreDOTNo?: any;
    RHRearTyreMake?: any;
    RHRearTyreDOTNo?: any;
    LHFrontTyreMake?: any;
    LHFrontTyreDOTNo?: any;
    LHRearTyreMake?: any;
    LHRearTyreDOTNo?: any;
    StepneyTypeMake?: any;
    StepneyTypeDOTNo?: any;
    Marketvalue?: any;
    InspectionDateRecommandedOn?: any;
    Cowl?: any;
    DashBoard?: any;
    LeftMudgaurd?: any;
    RightMudgaurd?: any;
    Cabin?: any;
    CabinLeftDoor?: any;
    CabinRightDoor?: any;
    LeftSideBody?: any;
    RightSideBody?: any;
    FrontBody?: any;
    RearBody?: any;
    WindowGlassesLS?: any;
    WindowGlassesRS?: any;
    ExcavatorCabinGlass?: any;
    CraneCabinGlass?: any;
    FrontExcavator?: any;
    CraneBucket?: any;
    CraneHook?: any;
    Boom?: any;
    Ac?: any;
    Fans?: any;
    HydraulicSystem?: any;
    ChassisFrame?: any;
    FuelTank?: any;
    Seats?: any;
    Stepny?: any;
    ChassisProductionNo?: any;
    TypeofBody?: any;
    ExtraFittingsRemarks?: any;
    VehicleRemarks?: any;
    QCRemarks?: any;
    SurveyCompletedOn?: any;
    Status?: number;
    CreatedOn?: string;
    UpdatedOn?: any;
    CreatedBy?: number;
    UpdatedBy?: number;
}